Split cli, config and util functions into separate includefiles. Add LICENSE file.
